refactor(bpv7): rename compare module to cmp
